work for milestone 2
Question 1 is answered, Question 2 is also answered. Maybe we could also find the percent change over time though for that zip and compare it to the percent change of other zip codes? Question 3 I had trouble on with the demographics, Q4 I left for Mikayla, and Q5 I had trouble with the month labels but think I know how to fix it.